Revert "[FEAT, REFACTOR] PlayerAttack 리팩토링, 피격 구현"
