+++ Assets/01.Scripts/Enemy/Enemy/Enemy.cs
@@ -1,48 +1,27 @@
 using System;
 using UnityEngine;
 
-[RequireComponent(typeof(EnemyAI), typeof(EnemyMovement), typeof(EnemyHealth))]
 public class Enemy : MonoBehaviour
 {
     [SerializeField] private EnemyInfos infos;
 
-    public static event Action<Enemy> onEnemyDied;
+    public static event Action<Enemy> OnEnemyDied;
     public EnemyInfos enemyInfos => infos;
-
-    public EnemyMovement EnemyMovement {get; private set; }
-    public EnemyAI EnemyAi {get; private set; }
-    public EnemyHealth EnemyHealth {get; private set; }
-    // public EnemyAttack EnemyAttack {get; private set; }
-    
-    private void Awake()
-    {
-        EnemyMovement = GetComponent<EnemyMovement>();
-        EnemyAi = GetComponent<EnemyAI>();
-        EnemyHealth = GetComponent<EnemyHealth>();
-        // EnemyAttack = GetComponent<EnemyAttack>();
-    }
-
     // Start is called once before the first execution of Update after the MonoBehaviour is created
     void OnEnable()
     {
-        Invoke(nameof(Die), 15f);
-        EnemyHealth.onDeath += Die;
+        Invoke(nameof(die), 15f);
     }
 
-    void OnDisable()
-    {
-        EnemyHealth.onDeath -= Die;
-    }
-    
     // Update is called once per frame
     void Update()
     {
         
     }
 
-    void Die()
+    void die()
     {
-        onEnemyDied?.Invoke(this);
+        OnEnemyDied?.Invoke(this);
         Destroy(this.gameObject);
     }
 }
