--- Assets/01.Scripts/Enemy/Enemy/EnemyHealth.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: b5230db458234820a5efb0d34787985c
-timeCreated: 1748852101
