--- Assets/01.Scripts/Enemy/Enemy/HitboxRelay.cs
@@ -1,16 +1,0 @@
-using System;
-using UnityEngine;
-
-public class HitboxRelay : MonoBehaviour
-{
-    private EnemyHealth enemyHealth;
-    private void Awake()
-    {
-        enemyHealth = GetComponentInParent<EnemyHealth>();
-    }
-
-    private void OnTriggerEnter(Collider other)
-    {
-        enemyHealth?.OnHitboxTriggerEnter(other);
-    }
-}
