+++ Assets/01.Scripts/IDamageable.cs
@@ -1,0 +1,7 @@
+using System.Numerics;
+
+public interface IDamageable
+{
+    void OnDamage(float damage, Vector3 hitPoint, Vector3 hitNormal);
+    void Die();
+}
