+++ Assets/01.Scripts/IDamageable.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: a8ab5e5d1ea6b4ac9b43e83a39099782
