+++ Assets/01.Scripts/LivingEntity.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: c1389ddc58a604343b48f21d2ebcaf5c
