--- Assets/01.Scripts/Player/Weapon.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: acd8afd8634f4b918c7377e46ab91b2b
-timeCreated: 1748875264
