--- Assets/02.Prefabs/Enemy/Enemy/01.Enemy/Laptop Variant.prefab
@@ -1,84 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!1001 &7113841285167943079
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    serializedVersion: 3
-    m_TransformParent: {fileID: 0}
-    m_Modifications:
-    - target: {fileID: 515471733011550504, guid: b3a3128e59ab14effb32d3e681e48ed2, type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 515471733011550504, guid: b3a3128e59ab14effb32d3e681e48ed2, type: 3}
-      propertyPath: m_LocalPosition.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 515471733011550504, guid: b3a3128e59ab14effb32d3e681e48ed2, type: 3}
-      propertyPath: m_LocalPosition.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 515471733011550504, guid: b3a3128e59ab14effb32d3e681e48ed2, type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 515471733011550504, guid: b3a3128e59ab14effb32d3e681e48ed2, type: 3}
-      propertyPath: m_LocalRotation.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 515471733011550504, guid: b3a3128e59ab14effb32d3e681e48ed2, type: 3}
-      propertyPath: m_LocalRotation.y
-      value: -0
-      objectReference: {fileID: 0}
-    - target: {fileID: 515471733011550504, guid: b3a3128e59ab14effb32d3e681e48ed2, type: 3}
-      propertyPath: m_LocalRotation.z
-      value: -0
-      objectReference: {fileID: 0}
-    - target: {fileID: 515471733011550504, guid: b3a3128e59ab14effb32d3e681e48ed2, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 515471733011550504, guid: b3a3128e59ab14effb32d3e681e48ed2, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 515471733011550504, guid: b3a3128e59ab14effb32d3e681e48ed2, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 607499987148880288, guid: b3a3128e59ab14effb32d3e681e48ed2, type: 3}
-      propertyPath: m_Enabled
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 895233791870159762, guid: b3a3128e59ab14effb32d3e681e48ed2, type: 3}
-      propertyPath: m_Name
-      value: Laptop Variant
-      objectReference: {fileID: 0}
-    m_RemovedComponents:
-    - {fileID: 217931439759727011, guid: b3a3128e59ab14effb32d3e681e48ed2, type: 3}
-    m_RemovedGameObjects: []
-    m_AddedGameObjects: []
-    m_AddedComponents:
-    - targetCorrespondingSourceObject: {fileID: 2466431158899843369, guid: b3a3128e59ab14effb32d3e681e48ed2, type: 3}
-      insertIndex: -1
-      addedObject: {fileID: 6334624134505613619}
-  m_SourcePrefab: {fileID: 100100000, guid: b3a3128e59ab14effb32d3e681e48ed2, type: 3}
---- !u!1 &4648830712506461326 stripped
-GameObject:
-  m_CorrespondingSourceObject: {fileID: 2466431158899843369, guid: b3a3128e59ab14effb32d3e681e48ed2, type: 3}
-  m_PrefabInstance: {fileID: 7113841285167943079}
-  m_PrefabAsset: {fileID: 0}
---- !u!114 &6334624134505613619
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 4648830712506461326}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: a54b543621f7e4bcf84fc74a8ba1b24b, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier:
