--- Assets/02.Prefabs/Enemy/Enemy/01.Enemy/Laptop Variant.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 9b27fc14825484fffb8558d4a33adcfc
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
